Create and use trait for id and time in entities.
