+++ src/Entity/Student.php
@@ -2,6 +2,8 @@
 
 namespace App\Entity;
 
+use App\Entity\Traits\EntityIdTrait;
+use App\Entity\Traits\EntityTimeTrait;
 use Doctrine\ORM\Mapping as ORM;
 
 /**
@@ -9,17 +11,8 @@ use Doctrine\ORM\Mapping as ORM;
  */
 class Student
 {
-    /**
-     * @ORM\Id()
-     * @ORM\GeneratedValue()
-     * @ORM\Column(type="integer")
-     */
-    private $id;
-
-    /**
-     * @ORM\Column(type="string", length=32)
-     */
-    private $uuid;
+    use EntityIdTrait;
+    use EntityTimeTrait;
 
     /**
      * @ORM\Column(type="string", length=32)
@@ -52,36 +45,9 @@ class Student
     private $seondMark;
 
     /**
-     * @ORM\Column(type="datetime_immutable")
-     */
-    private $createdDate;
-
-    /**
-     * @ORM\Column(type="datetime", nullable=true)
-     */
-    private $updatedDate;
-
-    /**
      * @ORM\ManyToOne(targetEntity="App\Entity\Classroom", inversedBy="students")
      */
     private $classroom;
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function getUuid(): ?string
-    {
-        return $this->uuid;
-    }
-
-    public function setUuid(string $uuid): self
-    {
-        $this->uuid = $uuid;
-
-        return $this;
-    }
 
     public function getFirstname(): ?string
     {
@@ -151,30 +117,6 @@ class Student
     public function setSeondMark(?float $seondMark): self
     {
         $this->seondMark = $seondMark;
-
-        return $this;
-    }
-
-    public function getCreatedDate(): ?\DateTimeImmutable
-    {
-        return $this->createdDate;
-    }
-
-    public function setCreatedDate(\DateTimeImmutable $createdDate): self
-    {
-        $this->createdDate = $createdDate;
-
-        return $this;
-    }
-
-    public function getUpdatedDate(): ?\DateTimeInterface
-    {
-        return $this->updatedDate;
-    }
-
-    public function setUpdatedDate(?\DateTimeInterface $updatedDate): self
-    {
-        $this->updatedDate = $updatedDate;
 
         return $this;
     }
