+++ src/Entity/Traits/EntityIdTrait.php
@@ -1,0 +1,46 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: training
+ * Date: 1/30/20
+ * Time: 12:27 PM
+ */
+
+namespace App\Entity\Traits;
+
+use Doctrine\ORM\Mapping as ORM;
+use Ramsey\Uuid\UuidInterface;
+
+trait EntityIdTrait
+{
+
+    /**
+     * @ORM\Id()
+     * @ORM\GeneratedValue()
+     * @ORM\Column(type="integer", options={"unsigned": true})
+     */
+    protected $id;
+
+
+    /**
+     * @ORM\Column(type="uuid", unique=true)
+     */
+    protected $uuid;
+
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+
+    public function getUuid(): ?UuidInterface
+    {
+        return $this->uuid;
+    }
+
+    public function setUuid(string $uuid): self
+    {
+        $this->uuid = $uuid;
+
+        return $this;
+    }
+}
