+++ src/Entity/Traits/EntityTimeTrait.php
@@ -1,0 +1,49 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: training
+ * Date: 1/30/20
+ * Time: 12:32 PM
+ */
+
+namespace App\Entity\Traits;
+
+
+trait EntityTimeTrait
+{
+
+    /**
+     * @ORM\Column(type="datetime_immutable")
+     */
+    protected $createdDate;
+
+    /**
+     * @ORM\Column(type="datetime", nullable=true)
+     */
+    protected $updatedDate;
+
+    public function getCreatedDate(): ?\DateTimeImmutable
+    {
+        return $this->createdDate;
+    }
+
+    public function setCreatedDate(\DateTimeImmutable $createdDate): self
+    {
+        $this->createdDate = $createdDate;
+
+        return $this;
+    }
+
+    public function getUpdatedDate(): ?\DateTimeInterface
+    {
+        return $this->updatedDate;
+    }
+
+    public function setUpdatedDate(?\DateTimeInterface $updatedDate): self
+    {
+        $this->updatedDate = $updatedDate;
+
+        return $this;
+    }
+
+}
